Add more tests for identity
