+++ contracts/CxipIdentity.sol
@@ -273,7 +273,7 @@ contract CxipIdentity {
             );
         }
         require(_isOwner(collectionCreator), "CXIP: creator not owner");
-        bytes memory bytecode = hex"608060405234801561001057600080fd5b5060f68061001f6000396000f3fe60806040819052632c5feccb60e11b815260009073ed51d2028f51dff94a6036e502caab0d6717ab03906358bfd99690608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea2646970667358221220c554917349a8d48b2187dbfaed113ce018db35db61483b684aa61d7e907cb1f764736f6c634300080c0033";
+        bytes memory bytecode = hex"608060405234801561001057600080fd5b5060f68061001f6000396000f3fe60806040819052632c5feccb60e11b815260009073e7f1725e7734ce288f8367e1bb143e90bb3f0512906358bfd99690608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea26469706673582212208e23a2c91da5b61060208c76459fbc6d3bb5dab07c387a6c4bd84d044985701364736f6c634300080c0033";
         address cxipAddress;
         assembly {
             cxipAddress := create2(
@@ -338,7 +338,7 @@ contract CxipIdentity {
             )
         }
         require(
-            keccak256(cxipAddress.code) == keccak256(ICxipRegistry(0xeD51D2028F51DFF94a6036E502caAB0d6717Ab03).getCustomSource(slot).code),
+            keccak256(cxipAddress.code) == keccak256(ICxipRegistry(0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512).getCustomSource(slot).code),
             "CXIP: byte code missmatch"
         );
         ICxipERC721(cxipAddress).init(collectionCreator, collectionData);
@@ -643,6 +643,6 @@ contract CxipIdentity {
      * @return ICxipRegistry The address of the top-level CXIP Registry smart contract.
      */
     function getRegistry() internal pure returns (ICxipRegistry) {
-        return ICxipRegistry(0xeD51D2028F51DFF94a6036E502caAB0d6717Ab03);
+        return ICxipRegistry(0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512);
     }
 }
