+++ contracts/CxipProvenance.sol
@@ -141,7 +141,7 @@ contract CxipProvenance {
                 "CXIP: invalid signature"
             );
         }
-        bytes memory bytecode = hex"608060405234801561001057600080fd5b5060f68061001f6000396000f3fe6080604081905263071b938d60e31b815260009073ed51d2028f51dff94a6036e502caab0d6717ab03906338dc9c6890608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea264697066735822122057a9f6abe4e224621c5924212f28abc02c6c072ee01cf42e724b6d60fcbde90e64736f6c634300080c0033";
+        bytes memory bytecode = hex"608060405234801561001057600080fd5b5060f68061001f6000396000f3fe6080604081905263071b938d60e31b815260009073e7f1725e7734ce288f8367e1bb143e90bb3f0512906338dc9c6890608490602090600481865afa158015604b573d6000803e3d6000fd5b505050506040513d601f19601f82011682018060405250810190606d91906092565b90503660008037600080366000845af43d6000803e808015608d573d6000f35b3d6000fd5b60006020828403121560a357600080fd5b81516001600160a01b038116811460b957600080fd5b939250505056fea26469706673582212208946e1a7de585e83ddba36332acc62598a9d62e903d96e35157330f329e2f70d64736f6c634300080c0033";
         address identityAddress;
         assembly {
             identityAddress := create2(
@@ -278,6 +278,6 @@ contract CxipProvenance {
      * @return ICxipRegistry The address of the top-level CXIP Registry smart contract.
      */
     function getRegistry() internal pure returns (ICxipRegistry) {
-        return ICxipRegistry(0xeD51D2028F51DFF94a6036E502caAB0d6717Ab03);
+        return ICxipRegistry(0xe7f1725E7734CE288F8367e1Bb143E90bb3F0512);
     }
 }
